Implement BeatmapHeaderContent
